Add new component for join and leave button
